Tidy up API test functions and test data out of test files.
