+++ playwright_tests/api/update-pba.negative.api.test.ts
@@ -1,37 +1,8 @@
 import { test, expect } from './helpers/api.fixtures';
 import { cleanupProvisionedOrganisation, provisionPendingOrganisation } from './helpers/organisations-write.helpers';
+import { UPDATE_PBA_MALFORMED_CASES } from './helpers/update-pba.helpers';
 
 const UNAUTHENTICATED_ORG_ID = 'UPDATE_PBA_NEGATIVE_ORG';
-const UPDATE_PBA_VALIDATION_ERROR_STATUSES = [400, 403, 404, 422, 500] as const;
-
-type UpdatePbaMalformedCase = {
-  name: string;
-  expectedStatuses: readonly number[];
-  buildPayload: (organisationId: string) => Record<string, unknown>;
-};
-
-const UPDATE_PBA_MALFORMED_CASES: UpdatePbaMalformedCase[] = [
-  {
-    name: 'empty request body',
-    buildPayload: () => ({}),
-    expectedStatuses: [400]
-  },
-  {
-    name: 'missing orgId',
-    buildPayload: () => ({ paymentAccounts: ['PBA33L6BNO'] }),
-    expectedStatuses: [400]
-  },
-  {
-    name: 'invalid orgId',
-    buildPayload: () => ({ paymentAccounts: ['PBA33L6BNO'], orgId: 'INVALID_ORG_ID_12345' }),
-    expectedStatuses: UPDATE_PBA_VALIDATION_ERROR_STATUSES
-  },
-  {
-    name: 'invalid paymentAccounts type',
-    buildPayload: (organisationId) => ({ paymentAccounts: 'PBA33L6BNO', orgId: organisationId }),
-    expectedStatuses: [400]
-  }
-];
 
 test.describe('Playwright API negative: update pba', { tag: ['@update-pba', '@negative'] }, () => {
   test('PUT /api/updatePba without auth is denied', async ({ apiAnonymousRequest }) => {
